refactor(quadraui): move lib.rs test dumping ground into primitive test modules (#510)
lib.rs was a ~320-line facade plus a 4,230-line #[cfg(test)] mod holding
163 tests for nearly every primitive, banner-sectioned but never moved
into the primitive files that already had their own test modules.

Mechanically relocate each section (and the un-bannered serde-roundtrip
tests that preceded the banners) into primitives/<name>.rs::tests (or
types.rs::tests for the non-primitive Color/WidgetId/StyledText tests),
appending to existing test modules where present and creating a fresh
`#[cfg(test)] mod tests` otherwise. Self-referential `primitives::<mod>::`
qualifications (needed from lib.rs's crate-root vantage point) are
stripped to bare names now that the code lives inside its own module.
Two genuine name collisions surfaced by the move are resolved:
`text_display.rs` had its own `make_td` helper with a different
signature (renamed the moved one to `make_td_from_lines`), and both
`form.rs` and `text_display.rs` needed a test-scoped `use
crate::types::Color;` since Color is only used from their moved tests.

Pure move, no behavior change: lib.rs shrinks from 4,624 to 394 lines
(facade + one crate-level test), and `cargo test --features tui --lib`
still reports exactly 1,874 passing (the pre-move baseline). Full
quality gate (tui/gtk/win legs + fmt) is green.

This lands bullet 1 of #510; bullets 2-4 (gtk/backend.rs layout
extraction, AppShell::handle decomposition, dispatch.rs/terminal_engine.rs
submodules) are independently landable follow-ups per the issue.
